refactor lyrics fetchers
